# Optimize sparse search by precomputing once the query info per posting