This PR is a low hanging fruit I noticed while working on a larger optimization.

We currently fetch the query weight corresponding to a posting list iterator every time we score a point.
This small cost adds up and actually shows in profiling.

The proposed fix is to precompute for each posting list their corresponding query information.

The result is a consistent 3% gain on the inverted index search:

```
sparse-vector-search-group/inverted-index-search
                        time:   [30.959 ms 31.025 ms 31.098 ms]
                        change: [-4.5531% -3.8959% -3.3699%] (p = 0.00 < 0.05)
                        Performance has improved.

sparse-vector-search-group/inverted-index-filtered-payload-index
                        time:   [34.407 ms 34.544 ms 34.693 ms]
                        change: [-3.3703% -2.7161% -2.1013%] (p = 0.00 < 0.05)
                        Performance has improved.
```

And  8% on the plain search:

```
sparse-vector-search-group/inverted-index-filtered-plain
                        time:   [268.72 ms 269.30 ms 269.89 ms]
                        change: [-8.7347% -8.4290% -8.1481%] (p = 0.00 < 0.05)
                        Performance has improved.

sparse-vector-search-group/plain-filtered-payload-index
                        time:   [260.24 ms 260.75 ms 261.30 ms]
                        change: [-8.0965% -7.8180% -7.5645%] (p = 0.00 < 0.05)
                        Performance has improved.
```